+++ Frameworks/Adium Framework/Source/CSNewContactAlertWindowController.m
@@ -39,7 +39,7 @@
 
 - (void)updateHeaderView;
 
-- (void)sheetDidEnd:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo;
+- (void)sheetDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
 - (void)alertDetailsForHeaderChanged:(NSNotification *)aNotification;
 
 @end
@@ -153,7 +153,7 @@
 }
 
 // Called as the user list edit sheet closes, dismisses the sheet
-- (void)sheetDidEnd:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo
+- (void)sheetDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo
 {
 	[self cleanUpDetailsPane];
